chore: add jsdoc, refactoring utils
