+++ jsdoc.json
@@ -1,0 +1,32 @@
+{
+  "source": {
+    "include": ["src", "package.json", "README.md"],
+    "includePattern": ".js$",
+    "excludePattern": "(node_modules/|docs)"
+  },
+  "plugins": ["plugins/markdown"],
+  "opts": {
+    "encoding": "utf8",
+    "readme": "./README.md",
+    "destination": "docs/",
+    "recurse": true,
+    "verbose": true,
+    "template": "./node_modules/clean-jsdoc-theme",
+    "theme_opts": {
+      "default_theme": "dark"
+    }
+  },
+  "markdown": {
+    "hardwrap": false,
+    "idInHeadings": true
+  },
+  "theme_opts": {
+    "homepageTitle": "PennyLaneTech - Документация Frontend CRM"
+  },
+  "meta": [
+    {
+      "name": "author",
+      "content": "PennyLaneTech"
+    }
+  ]
+}
