+++ package.json
@@ -9,7 +9,8 @@
     "stage": "vite build --mode stage",
     "lint": "eslit src",
     "dev": "vite",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "docs": "jsdoc --configure jsdoc.json --verbose"
   },
   "dependencies": {
     "@kyvg/vue3-notification": "^2.3.3",
@@ -43,6 +44,7 @@
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.0.4",
+    "clean-jsdoc-theme": "^4.3.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^7.20.0",
     "prettier": "^3.1.1",
