+++ src/components/Complex/Deal/Price/ComplexDealPriceStorage.vue
@@ -32,7 +32,7 @@
 import WithUnitType from '@/components/common/WithUnitType.vue';
 import { mapper } from '@/utils/mapper';
 import { entityProperties } from '@/const/properties/properties';
-import { reducer } from '@/utils';
+import { reducer } from '@/utils/reducer.js';
 import EmptyData from '@/components/common/EmptyData.vue';
 
 export default {
