+++ src/components/Forms/Complex/FormComplexBuildingCreate.vue
@@ -257,7 +257,7 @@
 <script>
 import { helpers, maxValue, minValue, required } from '@vuelidate/validators';
 import { mapActions } from 'vuex';
-import { yandexmap } from '@/utils';
+import { yandexmap } from '@/utils/yandexMap.js';
 import Loader from '@/components/common/Loader.vue';
 import { ComplexFormMixin } from '@/components/Forms/Complex/mixin';
 import Modal from '@/components/common/Modal.vue';
