+++ src/components/Forms/Equipment/FormEquipment.vue
@@ -201,7 +201,8 @@ import MultiSelect from '@/components/common/Forms/MultiSelect.vue';
 import useValidate from '@vuelidate/core';
 import { helpers, maxLength, minLength, minValue, required } from '@vuelidate/validators';
 import { mapGetters } from 'vuex';
-import { multiselectAdapter, optionsToList, yandexmap } from '@/utils';
+import { multiselectAdapter, optionsToList } from '@/utils';
+import { yandexmap } from '@/utils/yandexMap.js';
 import api from '@//api/api.js';
 import Modal from '@/components/common/Modal.vue';
 import CheckboxChip from '@/components/common/Forms/CheckboxChip.vue';
